requirements.txt
Django==1.10.3
certifi==2015.04.28
defusedxml==0.4.1
django-allauth==0.28.0
django-bootstrap-form==3.2.1
django-extensions
django-haystack==2.5.1
-e git+https://github.com/MasterGowen/django-tracking2@master
elasticsearch==1.9.0
mysqlclient==1.3.9
oauthlib==2.0.0
pyelasticsearch==1.4
python3-openid==3.0.10
pytz==2016.7
requests==2.12.1
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
urllib3==1.19.1
py-moneyed
django-money
django-phonenumber-field